remove old generic impl
