+++ polars/polars-lazy/polars-pipe/src/executors/sinks/groupby/string.rs
@@ -14,7 +14,7 @@ use polars_utils::unwrap::UnwrapUncheckedRelease;
 use rayon::prelude::*;
 
 use super::aggregates::AggregateFn;
-use super::generic::Key;
+use super::generic2::Key;
 use crate::executors::sinks::groupby::aggregates::AggregateFunction;
 use crate::executors::sinks::groupby::ooc_state::OocState;
 use crate::executors::sinks::groupby::physical_agg_to_logical;
